requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pypi_0
blobfile=2.0.1=pypi_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py38h06a4308_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=pypi_0
docker-pycreds=0.4.0=pypi_0
filelock=3.9.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
idna=3.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=4.9.2=pypi_0
mpi4py=3.1.4=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.24.2=pypi_0
openssl=1.1.1s=h7f8727e_0
pathtools=0.1.2=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py38h06a4308_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
pycryptodomex=3.17=pypi_0
python=3.8.16=h7a1cb2a_2
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
sentry-sdk=1.15.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
tk=8.6.12=h1ccaba5_0
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.4.0=pypi_0
urllib3=1.26.14=pypi_0
wandb=0.13.10=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
